further clean up code
