+++ experimental/rel_component/configs/rel_tok2vec.cfg
@@ -39,7 +39,6 @@ threshold = 0.5
 
 [components.relation_extractor.model]
 @architectures = "rel_model.v1"
-nO = null
 
 [components.relation_extractor.model.create_candidate_tensor]
 @misc = "rel_cand_tensor.v1"
