+++ experimental/rel_component/scripts/custom_functions.py
@@ -9,8 +9,7 @@ from spacy.tokens import DocBin, Doc
 from scripts.rel_pipe import make_relation_extractor
 
 # make the config work
-from scripts.rel_model import create_relation_model, create_classification_layer
-from scripts.rel_instances import create_candidates, create_tensors
+from scripts.rel_model import create_relation_model, create_classification_layer, create_candidates, create_tensors
 
 
 @registry.readers("Gold_ents_Corpus.v1")
